test(onboarding): Move the no-destination reuse case to the hook spec
The view spec was the only cover for a project created with Set up later and then submitted with a destination. Cover that branch of isSameSelection in the hook spec instead, where it needs no rendering.

Drop the repeated persist-together rationale from the option docstring and the hook spec, and let the platform-features spec overrides carry CreatedProject directly.

Refs VDY-141
